Do not delete if model object doesnot exist

+++ qshop/cart/cart.py
@@ -200,8 +200,9 @@ class Cart:
         else:
             item.delete()
         if self.total_products() == 0:
-            self.cart.delete()
-            del self.__request.session[CART_ID]
+            if self.cart.pk:
+                self.cart.delete()
+                del self.__request.session[CART_ID]
 
     def update(self, item_id, quantity):
         self.clear_cache()
